+++ Changelog.md
@@ -1,3 +1,6 @@
+## 0.0.32
+*INCOMPLETE*
+
 ## 0.0.31
 + Added experimental support for making on-device predictions.
 + Added `PredictionResource.name` field for handling prediction resources with required file names.